requirements.txt
numpy>=1.21.6
torch>=2.0.1+cu118
torchvision>=0.15.2+cu118
torchaudio>=2.0.2+cu118
opencv-python-headless>=4.1.2
easyocr>=1.6.2
streamlit>=1.22.0
Pillow>=7.1.2
PyMuPDF>=1.22.2
google-generativeai>=0.0.3
